backend/app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
pydantic
httpx